pyproject.toml
[tool.poetry]
name = "exercise_dm"
version = "0.1.0"
description = "Semantic similarity with context"
authors = ["Yaxiong Yuan <yyx1364@gmail.com>"]
readme = "README.md"
packages = [{ include = "exercise_dm", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
torch = { url = "https://download.pytorch.org/whl/cpu/torch-2.3.0-cp311-none-macosx_11_0_arm64.whl#sha256=d24e328226d8e2af7cf80fcb1d2f1d108e0de32777fab4aaa2b37b9765d8be73" }
sentence-transformers = { version = "3.4.1", extras = ["train"] }
scikit-learn = "1.6.1"
datasets = "3.2.0"
fastapi = "0.115.8"
uvicorn = "0.34.0"

[tool.poetry.dev-dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core"]